+++ bn-apps/cordapp-updates-distribution/corda-updates-app/src/main/kotlin/net/corda/cordaupdates/app/member/SyncerService.kt
@@ -3,7 +3,6 @@ package net.corda.cordaupdates.app.member
 import net.corda.businessnetworks.cordaupdates.core.ArtifactMetadata
 import net.corda.businessnetworks.cordaupdates.core.CordappSyncer
 import net.corda.businessnetworks.cordaupdates.core.SyncerConfiguration
-import net.corda.cordaupdates.transport.SessionConfigurationProperties
 import net.corda.core.node.AppServiceHub
 import net.corda.core.node.services.CordaService
 import net.corda.core.serialization.SingletonSerializeAsToken
@@ -54,8 +53,6 @@ internal class SyncerService(private val appServiceHub : AppServiceHub) : Single
                 CordappSyncer(SyncerConfiguration.readFromFile(config))
             } else CordappSyncer(syncerConfiguration)
 
-    private fun additionalConfigurationProperties() = mapOf(Pair(SessionConfigurationProperties.APP_SERVICE_HUB, appServiceHub))
-
     private fun refreshMetadataCacheSynchronously(syncerConfiguration : SyncerConfiguration? = null, body : (syncer : CordappSyncer) -> List<ArtifactMetadata>) : List<ArtifactMetadata> {
         val artifacts : List<ArtifactMetadata> = body(syncer(syncerConfiguration))
         val artifactsMetadataCache : ArtifactsMetadataCache = appServiceHub.cordaService(ArtifactsMetadataCache::class.java)
@@ -67,7 +64,7 @@ internal class SyncerService(private val appServiceHub : AppServiceHub) : Single
      * Synchronously launches artifact synchronisation.
      */
     fun syncArtifacts(syncerConfiguration : SyncerConfiguration? = null) = refreshMetadataCacheSynchronously(syncerConfiguration) { syncer ->
-        syncer.syncCordapps(additionalConfigurationProperties = additionalConfigurationProperties())
+        syncer.syncCordapps()
     }
 
     /**
@@ -76,7 +73,7 @@ internal class SyncerService(private val appServiceHub : AppServiceHub) : Single
      * @param cordappCoordinatesWithRange coordinates of a cordapp with range, i.e. "net.corda:corda-finance:[0,2.0)"
      */
     fun getArtifactsMetadata(cordappCoordinatesWithRange : String, syncerConfiguration : SyncerConfiguration? = null) = refreshMetadataCacheSynchronously(syncerConfiguration) { syncer ->
-        syncer.getAvailableVersions(cordappCoordinatesWithRange, additionalConfigurationProperties())
+        syncer.getAvailableVersions(cordappCoordinatesWithRange)
     }
 
     /**
