+++ bn-apps/cordapp-updates-distribution/corda-updates-core/src/test/kotlin/net/corda/businessnetworks/cordaupdates/core/CordappSyncerTest.kt
@@ -50,10 +50,6 @@ class CordappSyncerTest {
                 httpProxyPort = 187,
                 httpProxyUsername = "ProxyUsername",
                 httpProxyPassword = "ProxyPassword",
-                rpcHost = "localhost",
-                rpcPort = 8007,
-                rpcUsername = "RpcUsername",
-                rpcPassword = "RpcPassword",
                 cordappSources = listOf(
                         CordappSource(
                                 remoteRepoUrl = "http://search.maven.org",
