--- bn-apps/cordapp-updates-distribution/corda-updates-shell/src/main/resources/log4j2.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Configuration status="info">
-    <Properties>
-        <Property name="consoleLogLevel">off</Property>
-    </Properties>
-    <Appenders>
-        <Console name="STDOUT" target="SYSTEM_OUT" ignoreExceptions="false">
-            <PatternLayout pattern="[%C{1}.%M] %m%n"/>
-        </Console>
-    </Appenders>
-    <Loggers>
-        <Root level="${sys:consoleLogLevel}">
-            <AppenderRef ref="STDOUT"/>
-        </Root>
-    </Loggers>
-</Configuration>
