--- bn-apps/cordapp-updates-distribution/corda-updates-shell/src/main/resources/settings.mustache
@@ -1,46 +1,0 @@
-# Path to the local repository
-localRepoPath = "{{localRepoPath}}"
-
-# HTTP or HTTPS. Should be specified only if HTTP(S) proxy is used.
-# httpProxyType = "HTTP"
-
-# Proxy host name. Should be specified only if HTTP(S) proxy is used.
-# httpProxyHost = "10.0.0.1"
-
-# Proxy port. Should be specified only if HTTP(S) proxy is used.
-# httpProxyPort = "10009"
-
-# Username for proxy authentication. Should be specified only if HTTP(S) proxy is used.
-# httpProxyUsername = "proxy_user"
-
-# Password for proxy authentication. Should be specified only if HTTP(S) proxy is used.
-# httpProxyPassword = "P@$$w0rD"
-
-# RPC host of a Corda node to connect to. Should be specified only if corda-rpc or corda-auto transport is used.
-# rpcHost = "localhost"
-
-# RPC port of a Corda node to connect to. Should be specified only if corda-rpc or corda-auto transport is used.
-# rpcPort = "8003"
-
-# RPC username of a Corda node to connect to. Should be specified only if corda-rpc or corda-auto transport is used.
-# rpcUsername = "johndoe"
-
-# RPC password of a Corda node to connect to. Should be specified only if corda-rpc or corda-auto transport is used.
-# rpcPassword = "10004"
-
-# List of CorDapps to sync from remote repos. Can sync multiple CorDapps from multiple remote repositories
-cordappSources = [
-    {
-        # URL of the remote repository to fetch the cordapps from
-        remoteRepoUrl = "https://repo.maven.apache.org/maven2/"
-
-        # List of the cordapps to sync from the remote repository. Should be specified in the form of "artifactGroup:artifactName"
-        cordapps = ["net.corda:corda-finance"]
-
-        # Username, if the remote repo requires basic HTTP authentication
-        # httpUsername = "repo_user"
-
-        # Password, if the remote repo requires basic HTTP authentication
-        # httpPassword = "r3p0_P@$$"
-    }
-]
