--- bn-apps/cordapp-updates-distribution/corda-updates-shell/src/test/kotlin/net/corda/businessnetworks/cordaupdates/shell/CordaUpdatesCLITest.kt
@@ -1,4 +1,0 @@
-package net.corda.businessnetworks.cordaupdates.shell
-
-// TODO: write tests
-class CordaUpdatesCLITest
