--- bn-apps/cordapp-updates-distribution/corda-updates-transport/build.gradle
@@ -1,60 +1,0 @@
-apply plugin: 'kotlin'
-apply plugin: 'net.corda.plugins.cordapp'
-apply plugin: 'net.corda.plugins.quasar-utils'
-
-cordapp {
-    info {
-        name "corda-updates-transport"
-        vendor "com.r3"
-    }
-}
-
-sourceSets {
-    main {
-        resources {
-            srcDir "config/dev"
-        }
-    }
-    test {
-        resources {
-            srcDir "config/test"
-        }
-    }
-}
-
-dependencies {
-    compile "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
-    // maven resolver dependencies
-    compile "org.apache.maven.resolver:maven-resolver-api:$maven_resolver_version"
-    compile "org.apache.maven.resolver:maven-resolver-impl:$maven_resolver_version"
-    compile "org.apache.maven.resolver:maven-resolver-connector-basic:$maven_resolver_version"
-    compile "org.apache.maven.resolver:maven-resolver-transport-http:$maven_resolver_version"
-    compile "org.apache.maven.resolver:maven-resolver-transport-file:$maven_resolver_version"
-    compile "org.apache.maven.resolver:maven-resolver-util:$maven_resolver_version"
-    compile "org.apache.maven.resolver:maven-resolver-spi:$maven_resolver_version"
-    compile("org.apache.maven:maven-resolver-provider:3.5.4") {
-        exclude group: "org.apache.maven.resolver", module: "*"
-    }
-    // corda dependencies
-    cordaCompile "$corda_release_group:corda-core:$corda_release_version"
-    cordaCompile "$corda_release_group:corda-node-api:$corda_release_version"
-    cordaCompile "$corda_release_group:corda:$corda_release_version"
-    compile "$corda_release_group:corda-rpc:$corda_release_version"
-    testCompile "$corda_release_group:corda-node-driver:$corda_release_version"
-    testCompile project(":bn-apps:cordapp-updates-distribution:corda-updates-test-utils")
-
-
-    // test dependencies
-    testCompile "org.jetbrains.kotlin:kotlin-test:$kotlin_version"
-    testCompile "junit:junit:$junit_version"
-    testCompile project(':bn-apps:cordapp-updates-distribution:corda-updates-core')
-}
-
-tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
-    kotlinOptions {
-        languageVersion = "1.2"
-        apiVersion = "1.2"
-        jvmTarget = "1.8"
-        javaParameters = true   // Useful for reflection.
-    }
-}
