--- bn-apps/cordapp-updates-distribution/corda-updates-transport/src/test/kotlin/net/corda/cordaupdates/transport/SessionFilters.kt
@@ -1,13 +1,0 @@
-package net.corda.cordaupdates.transport
-
-import net.corda.cordaupdates.transport.flows.SessionFilter
-import net.corda.core.flows.FlowLogic
-import net.corda.core.flows.FlowSession
-
-class DenyAllSessionFilter : SessionFilter {
-    override fun isSessionAllowed(session : FlowSession, flowLogic : FlowLogic<*>) = false
-}
-
-class AllowAllSessionFilter : SessionFilter {
-    override fun isSessionAllowed(session : FlowSession, flowLogic : FlowLogic<*>) = true
-}
