--- bn-apps/cordapp-updates-distribution/corda-updates-transport/src/test/resources/corda-updates-app.conf
@@ -1,1 +1,0 @@
-remoteRepoUrl="file:../TestRepo"
