--- bn-apps/cordapp-updates-distribution/corda-updates-transport/src/test/resources/corda-updates-with-allow-filter.conf
@@ -1,2 +1,0 @@
-remoteRepoUrl="file:../TestRepo"
-sessionFilter=net.corda.cordaupdates.transport.AllowAllSessionFilter
