--- bn-apps/cordapp-updates-distribution/corda-updates-transport/src/test/resources/corda-updates-with-deny-filter.conf
@@ -1,2 +1,0 @@
-remoteRepoUrl="file:../TestRepo"
-sessionFilter=net.corda.cordaupdates.transport.DenyAllSessionFilter
